validation & dynamic timeline
